add Payment and Deli

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/OrderDto.java
@@ -28,9 +28,9 @@ public class OrderDto {
 
     private Boolean fragile;
 
-    private Integer totalPrice;
+    private Double totalPrice;
 
-    private Integer deliveryPrice;
+    private Double deliveryPrice;
 
-    private Integer productPrice;
+    private Double productPrice;
 }
